pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spiritual-transcript-pipeline"
version = "0.1.0"
description = "CLI pipeline for timestamped transcription, LLM-based module segmentation, and video clipping."
requires-python = ">=3.10"
dependencies = [
  "langchain-core>=0.3.0",
  "langchain-openai>=0.2.0",
  "langchain-anthropic>=0.3.0",
  "langchain-google-genai>=2.0.0",
  "langgraph>=0.2.0",
  "openai>=1.40.0",
  "python-dotenv>=1.0.1",
]

[project.scripts]
stp = "spiritual_transcript_pipeline.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
